requirements.txt
fastapi
uvicorn
pydantic
pandas
google-generativeai
python-multipart
aiofiles
jinja2
requests
